# Improve performance/memory usage of HashJoin datastructure (5-15% improvement on selected TPC-H queries)