… datafusion/core/src/physical_plan/joins/symmetric_hash_join.rs