# Which issue does this PR close?

Closes #https://github.com/apache/arrow-datafusion/issues/6700

<!--
We generally require a GitHub issue to be filed for all bug fixes and enhancements and this helps us generate change logs for our releases. You can link an issue to this PR using the GitHub syntax. For example `Closes #123` indicates that this PR will close issue #123.
-->

Benchmark results (TPC-H SF=1 in memory, average of 20 runs): q5, q7, q17, q18, and q21 show some improvement, most queries do not challenge the join or this code path (building the hashmap) of the join too much.

|  query | main   |  PR  |
|---|---|---|
| 5 | 58.74  | 54.98  |
|  7 |  104.14 | 92.41  |
|  17 | 332.83  |  320.09 |
|  18 | 244.41  | 218.97  |
|  21  | 152.79  | 132.93  |


Using the bench script (tpch_mem):
```
┏━━━━━━━━━━━━━━┳━━━━━━━━━━┳━━━━━━━━━━━━━━━━━━━━━┳━━━━━━━━━━━━━━━┓
┃ Query        ┃     main ┃ adapt_datastructure ┃        Change ┃
┡━━━━━━━━━━━━━━╇━━━━━━━━━━╇━━━━━━━━━━━━━━━━━━━━━╇━━━━━━━━━━━━━━━┩
│ QQuery 1     │ 187.78ms │            188.39ms │     no change │
│ QQuery 2     │  70.28ms │             68.41ms │     no change │
│ QQuery 3     │  52.51ms │             53.41ms │     no change │
│ QQuery 4     │  39.13ms │             38.95ms │     no change │
│ QQuery 5     │ 126.46ms │            113.56ms │ +1.11x faster │
│ QQuery 6     │  11.12ms │             11.14ms │     no change │
│ QQuery 7     │ 277.91ms │            232.85ms │ +1.19x faster │
│ QQuery 8     │  80.91ms │             81.30ms │     no change │
│ QQuery 9     │ 174.15ms │            163.58ms │ +1.06x faster │
│ QQuery 10    │ 104.81ms │            104.20ms │     no change │
│ QQuery 11    │  53.20ms │             54.47ms │     no change │
│ QQuery 12    │  70.71ms │             71.59ms │     no change │
│ QQuery 13    │ 214.11ms │            202.77ms │ +1.06x faster │
│ QQuery 14    │  14.09ms │             13.13ms │ +1.07x faster │
│ QQuery 15    │  23.20ms │             23.99ms │     no change │
│ QQuery 16    │  67.71ms │             51.89ms │ +1.30x faster │
│ QQuery 17    │ 718.41ms │            705.77ms │     no change │
│ QQuery 18    │ 732.54ms │            607.96ms │ +1.20x faster │
│ QQuery 19    │  61.01ms │             61.89ms │     no change │
│ QQuery 20    │ 214.62ms │            218.56ms │     no change │
│ QQuery 21    │ 412.88ms │            335.48ms │ +1.23x faster │
│ QQuery 22    │  33.44ms │             34.58ms │     no change │
└──────────────┴──────────┴─────────────────────┴───────────────┘
```

# Rationale for this change

Currently, we're using a `smallvec` to keep the indices with the same value. This hurts performance for cases when we have multiple values as it slows down insertion as well as retrieval of the indices (if the smallvec allocates, it's not cache-efficient).

Also the memory usage of the in-memory structure is reduced by paying the `Vec`-overhead only once rather than per-value: required memory is roughly reduced by 4x for storing the indices (8 bytes per value vs 32 bytes)

We can use a chained-list datastructure in a preallocated array where the next indices are stored (adapted from [Balancing vectorized query execution with bandwidth-optimized storage](https://dare.uva.nl/search?identifier=5ccbb60a-38b8-4eeb-858a-e7735dd37487))

<!--
 Why are you proposing this change? If this is already explained clearly in the issue then this section is not needed.
 Explaining clearly why changes are proposed helps reviewers understand your changes and offer better suggestions for fixes.  
-->

# What changes are included in this PR?

<!--
There is no need to duplicate the description in the issue here but it is sometimes worth providing a summary of the individual changes in this PR.
-->

# Are these changes tested?
Yes, by existing tests.

<!--
We typically require tests for all PRs in order to:
1. Prevent the code from being accidentally broken by subsequent changes
2. Serve as another way to document the expected behavior of the code

If tests are not included in your PR, please explain why (for example, are they covered by existing tests)?
-->

# Are there any user-facing changes?
No

<!--
If there are user-facing changes then we may require documentation to be updated before approving the PR.
-->

<!--
If there are any breaking changes to public APIs, please add the `api change` label.
-->