@Dandandan 2023-06-16
FYI @berkaysynnada I am trying to make some improvements to the hash join datastructure.

It seems the datastructure changes are not really compatible with the symmetric hash join (as it needs to be mutated during the probing process) - does it make sense to "duplicate" the old structure and code and use that inside the symmetric hash join?