@alamb 2023-06-18
Here are my measurements which I think are consistent with what is on this PR.
This is very exciting to see


```
--------------------
Benchmark tpch_mem.json
--------------------
┏━━━━━━━━━━━━━━┳━━━━━━━━━━━┳━━━━━━━━━━━━━━━━━━━━━┳━━━━━━━━━━━━━━━┓
┃ Query        ┃ main_base ┃ adapt_datastructure ┃        Change ┃
┡━━━━━━━━━━━━━━╇━━━━━━━━━━━╇━━━━━━━━━━━━━━━━━━━━━╇━━━━━━━━━━━━━━━┩
│ QQuery 1     │  517.86ms │            827.70ms │  1.60x slower │
│ QQuery 2     │  278.64ms │            534.33ms │  1.92x slower │
│ QQuery 3     │  186.82ms │            178.34ms │     no change │
│ QQuery 4     │  114.30ms │            112.48ms │     no change │
│ QQuery 5     │  480.70ms │            436.67ms │ +1.10x faster │
│ QQuery 6     │   40.20ms │             40.14ms │     no change │
│ QQuery 7     │ 1212.07ms │            963.47ms │ +1.26x faster │
│ QQuery 8     │  248.13ms │            251.24ms │     no change │
│ QQuery 9     │  613.36ms │            614.93ms │     no change │
│ QQuery 10    │  344.36ms │            339.00ms │     no change │
│ QQuery 11    │  212.06ms │            211.54ms │     no change │
│ QQuery 12    │  165.50ms │            167.81ms │     no change │
│ QQuery 13    │  673.03ms │            697.19ms │     no change │
│ QQuery 14    │   53.17ms │             50.71ms │     no change │
│ QQuery 15    │   88.08ms │             96.08ms │  1.09x slower │
│ QQuery 16    │  251.78ms │            200.89ms │ +1.25x faster │
│ QQuery 17    │ 2692.16ms │           2694.20ms │     no change │
│ QQuery 18    │ 2814.05ms │           2642.08ms │ +1.07x faster │
│ QQuery 19    │  167.80ms │            167.04ms │     no change │
│ QQuery 20    │  865.79ms │            852.77ms │     no change │
│ QQuery 21    │ 1446.79ms │           1268.84ms │ +1.14x faster │
│ QQuery 22    │  100.54ms │            102.31ms │     no change │
└──────────────┴───────────┴─────────────────────┴───────────────┘
```