@Dandandan 2023-06-18
> 

Thanks for running the benchmarks @alamb  - are you sure of the accuracy of the slower running queries?