@Dandandan 2023-06-18
Especially query 1 is suspicious, as it doesn't have a join @alamb ;)