@Dandandan 2023-06-18
> > Especially query 1 is suspicious, as it doesn't have a join @alamb ;)
> 
> Wow. So there is a huge noise in the benchmark?

I ran the benchmark manually, averaged over 20 runs, which has only minimal noise. But running once might add some more noise