@Dandandan 2023-06-19
My results using the bench script with default parameters:

```
┏━━━━━━━━━━━━━━┳━━━━━━━━━━┳━━━━━━━━━━━━━━━━━━━━━┳━━━━━━━━━━━━━━━┓
┃ Query        ┃     main ┃ adapt_datastructure ┃        Change ┃
┡━━━━━━━━━━━━━━╇━━━━━━━━━━╇━━━━━━━━━━━━━━━━━━━━━╇━━━━━━━━━━━━━━━┩
│ QQuery 1     │ 187.78ms │            188.39ms │     no change │
│ QQuery 2     │  70.28ms │             68.41ms │     no change │
│ QQuery 3     │  52.51ms │             53.41ms │     no change │
│ QQuery 4     │  39.13ms │             38.95ms │     no change │
│ QQuery 5     │ 126.46ms │            113.56ms │ +1.11x faster │
│ QQuery 6     │  11.12ms │             11.14ms │     no change │
│ QQuery 7     │ 277.91ms │            232.85ms │ +1.19x faster │
│ QQuery 8     │  80.91ms │             81.30ms │     no change │
│ QQuery 9     │ 174.15ms │            163.58ms │ +1.06x faster │
│ QQuery 10    │ 104.81ms │            104.20ms │     no change │
│ QQuery 11    │  53.20ms │             54.47ms │     no change │
│ QQuery 12    │  70.71ms │             71.59ms │     no change │
│ QQuery 13    │ 214.11ms │            202.77ms │ +1.06x faster │
│ QQuery 14    │  14.09ms │             13.13ms │ +1.07x faster │
│ QQuery 15    │  23.20ms │             23.99ms │     no change │
│ QQuery 16    │  67.71ms │             51.89ms │ +1.30x faster │
│ QQuery 17    │ 718.41ms │            705.77ms │     no change │
│ QQuery 18    │ 732.54ms │            607.96ms │ +1.20x faster │
│ QQuery 19    │  61.01ms │             61.89ms │     no change │
│ QQuery 20    │ 214.62ms │            218.56ms │     no change │
│ QQuery 21    │ 412.88ms │            335.48ms │ +1.23x faster │
│ QQuery 22    │  33.44ms │             34.58ms │     no change │
└──────────────┴──────────┴─────────────────────┴───────────────┘
```