@Dandandan 2023-06-19
> Currently, I am reviewing the code. The algorithm is neat, however, I want to find a good way to integrate the symmetric hash join into this algorithm. I think it can be possible.

@ozankabak I am not too familiar with the symmetric hash join, but one complexity seemed to be in reallocating/updating the `next` datastructure for the symmetric hash join.