@metesynnada 2023-06-19
I think changing the chain start pointer and a deletion offset might be the solution, I am working on that. We can address this in another PR. Since the performance gain is obvious, we can move on with this algorithm.

If you change the memory reservation calculation, LGTM.