@Dandandan 2023-06-19
> ```rust
> 32 bytes per `(u64, SmallVec<[u64; 1]>)
> ```

Sounds good :). Yes will do!