@alamb 2023-06-19
>  Thanks for running the benchmarks @alamb - are you sure of the accuracy of the slower running queries?

No I am not -- I observe significant variation on the queries that take small amounts of time to run. Thank you